fix:raven_channel_member apis
