+++ raven/raven_channel_management/doctype/raven_channel_member/raven_channel_member.py
@@ -3,8 +3,6 @@
 
 import frappe
 from frappe.model.document import Document
-from pypika import JoinType, Order
-from raven.api.raven_users import get_list
 
 class RavenChannelMember(Document):
 
@@ -67,17 +65,3 @@ class RavenChannelMember(Document):
 
     def get_admin_count(self):
         return frappe.db.count("Raven Channel Member", {"channel_id": self.channel_id, "is_admin": 1})
-
-@frappe.whitelist()
-def remove_channel_member(user_id, channel_id):
-    # Get raven channel member name where user_id and channel_id match
-    member = frappe.db.get_value("Raven Channel Member", {
-                                 "user_id": user_id, "channel_id": channel_id}, ["name"])
-    # Delete raven channel member
-    if member:
-        frappe.delete_doc("Raven Channel Member", member)
-        frappe.db.commit()
-    else: 
-        frappe.throw("User is not a member of this channel")
-    
-    return True
